refactor(cli): reorganise remote testbed into cli/src/remote/ module
Split the monolithic commands/remote_testbed.rs into a coherent module
hierarchy:

  commands/remote_testbed.rs  ← thin entry point (symmetry with other cmds)
  remote.rs                   ← RemoteTestbedDriver<C>: lifecycle + run()
  remote/benchmark.rs         ← RemoteBenchmarkDriver: benchmark + helpers
  remote/testbed.rs           ← TestbedStatusRender extension trait
  remote/protocol.rs          ← ReplicaProtocol, NodeParameters, ClientParameters
                                 (moved from cli/src/protocol.rs)

RemoteTestbedDriver no longer holds an ssh_manager — it is built in the
Benchmark arm of run() and passed to RemoteBenchmarkDriver, keeping the
lifecycle driver lean. RemoteBenchmarkDriver owns its own Progress instance
(single-use, one benchmark run per driver).
