+++ crates/cli/src/lib.rs
@@ -4,6 +4,6 @@
 pub mod args;
 pub mod commands;
 pub mod exporter;
-pub mod protocol;
+pub mod remote;
 pub mod terminal;
 pub mod tracing;
